+++ src/vec2/index.ts
@@ -1,0 +1,3 @@
+export * from './vec2'
+export * from './operator'
+export * from './compare'
